+++ test/bcdd/result/contract/nil_as_valid_value_checking_test.rb
@@ -5,53 +5,57 @@ require 'test_helper'
 class BCDD::Result
   class Contract::NiltAsValidValueCheckingTest < Minitest::Test
     test 'BCDD::Result::Expectations' do
-      _Result = BCDD::Result::Expectations.new(
-        success: {
-          ok: ->(value) {
-            case value
-            in Numeric then nil
-            end
-          }
+      contract = {
+        ok: ->(value) {
+          case value
+          in Numeric then nil
+          end
         }
-      )
+      }
+
+      _Result1 = BCDD::Result::Expectations.new(success: contract)
 
       assert_raises(Contract::Error::UnexpectedValue) do
-        _Result::Success(:ok, 1)
+        _Result1::Success(:ok, 1)
       end
 
-      Contract.nil_as_valid_value_checking!
+      BCDD::Result.config.pattern_matching.enable!(:nil_as_valid_value_checking)
 
-      result = _Result::Success(:ok, 1)
+      _Result2 = BCDD::Result::Expectations.new(success: contract)
+
+      result = _Result2::Success(:ok, 1)
 
       assert result.success?(:ok)
       assert_equal(1, result.value)
     ensure
-      Contract.nil_as_valid_value_checking!(enabled: false)
+      BCDD::Result.config.pattern_matching.disable!(:nil_as_valid_value_checking)
     end
 
     test 'BCDD::Result::Context::Expectations' do
-      _Result = BCDD::Result::Context::Expectations.new(
-        success: {
-          ok: ->(value) {
-            case value
-            in { number: Numeric } then nil
-            end
-          }
+      contract = {
+        ok: ->(value) {
+          case value
+          in { number: Numeric } then nil
+          end
         }
-      )
+      }
+
+      _Result1 = BCDD::Result::Context::Expectations.new(success: contract)
 
       assert_raises(Contract::Error::UnexpectedValue) do
-        _Result::Success(:ok, number: 1)
+        _Result1::Success(:ok, number: 1)
       end
 
-      Contract.nil_as_valid_value_checking!
+      BCDD::Result.config.pattern_matching.enable!(:nil_as_valid_value_checking)
 
-      result = _Result::Success(:ok, number: 1)
+      _Result2 = BCDD::Result::Context::Expectations.new(success: contract)
+
+      result = _Result2::Success(:ok, number: 1)
 
       assert result.success?(:ok)
       assert_equal({ number: 1 }, result.value)
     ensure
-      Contract.nil_as_valid_value_checking!(enabled: false)
+      BCDD::Result.config.pattern_matching.disable!(:nil_as_valid_value_checking)
     end
   end
 end